# swagger.yml
swagger: '2.0'
info:
  description: Http proxy API
  version: 1.0.0
  title: Http proxy API
  contact:
    email: duduardo23@gmail.com
paths:
  /proxy-admin/config:
    put:
      tags:
      - Proxy config
      summary: Update proxy configurations
      operationId: updateProxyConfig
      description: |
        This operation will be update proxy setting
      produces:
      - application/json
      parameters:
      - in: body
        name: proxy configuration
        schema:
          $ref: '#/definitions/ProxyConfig'
      responses:
        200:
          description: current proxy configuration
          schema:
            $ref: '#/definitions/ProxyConfig'
    get:
      tags:
      - Proxy config
      summary: Get current proxy configuration
      operationId: getProxyConfig
      description: This operation return the current proxy configuration
      produces:
      - application/json
      responses:
        200:
          description: current proxy configuration
          schema:
            $ref: '#/definitions/ProxyConfig'
...
definitions:
  ProxyConfig:
    type: object
    properties:
      routes:
        description: List of routes configurations
        type: array
        items:
          $ref: '#/definitions/RouteConfig'
      limits:
        description: List of rate limit rules
        type: array
        items:
          $ref: '#/definitions/LimitConfig'
  RouteConfig:
    description: Route configuration
    type: object
    required: 
      - pattern
      - backends
    properties:
      pattern:
        type: string
        example: /backend1
      backends:
        type: array
        items:
          type: string
        example: ["http://backend2a:3001", "http://backend2a:3002"]
  LimitConfig:
    description: Rate limite configuration
    type: object
    required: 
      - id
      - requestMin
    properties:
      id:
        type: string
        description: Unique rule ID
        example: rule01
      requestMin:
        type: integer
        description: Number of allowed request per minute
        example: 10000
      targetPath:
        type: string
        description: Url path that the rule must be applied.
        example: /backend2
      sourceIp:
        type: boolean
        description: Apply rule by request source ip
      headerValue:
        type: string
        description: Apply rule by request header value
        example: X-Apikey
...